% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraperFunctions.R
\name{activityCheckLinks}
\alias{activityCheckLinks}
\title{Scrapes all posts from an AC page}
\usage{
activityCheckLinks()
}
\value{
Returns a vector of all the AC links
}
\description{
Scrapes all posts from an AC page
}
