% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskScrapers.R
\name{activityCheckPosts}
\alias{activityCheckPosts}
\title{Scrapes all ACs and}
\usage{
activityCheckPosts(AC)
}
\arguments{
\item{AC}{A link to a specific AC page}
}
\value{
Returns a data frame of all the AC posts, by whom and when
}
\description{
Scrapes all ACs and
}
