% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraperFunctions.R
\name{affiliateLinks}
\alias{affiliateLinks}
\title{Scrapes all posts from an Affiliate page}
\usage{
affiliateLinks()
}
\value{
Returns a vector of all the Affiliate links
}
\description{
Scrapes all posts from an Affiliate page
}
