% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskScrapers.R
\name{affiliatePosts}
\alias{affiliatePosts}
\title{Scrapes all Affiliates and}
\usage{
affiliatePosts(Affiliate)
}
\arguments{
\item{Affiliate}{A link to a specific Affiliate page}
}
\value{
Returns a data frame of all the Affiliate posts, by whom and when
}
\description{
Scrapes all Affiliates and
}
