% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraperFunctions.R
\name{claimThreadPost}
\alias{claimThreadPost}
\title{Scrapes the claim threads}
\usage{
claimThreadPost(thread)
}
\arguments{
\item{thread}{A link to the claim thread}
}
\value{
Returns a data frame of all the AC posts, by whom and when
}
\description{
Scrapes the claim threads
}
