% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraperFunctions.R
\name{playerLinkScraper}
\alias{playerLinkScraper}
\title{Scrapes player links from teams}
\usage{
playerLinkScraper(forum)
}
\arguments{
\item{forum}{The forum link to the team}
}
\description{
Scrapes player links from teams
}
