% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraperFunctions.R
\name{playerScraper}
\alias{playerScraper}
\title{Scrapes all player page data}
\usage{
playerScraper(player)
}
\arguments{
\item{player}{Player page link}
}
\value{
Returns a data.frame with all information from a player page
}
\description{
Scrapes all player page data
}
