% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{readAPI}
\alias{readAPI}
\title{Calls the SSL API}
\usage{
readAPI(url, ...)
}
\arguments{
\item{url}{The url to the SSL API.}

\item{...}{Additional queries for the API call.}
}
\value{
A data frame of the data.
}
\description{
Calls the SSL API
}
\examples{
readAPI(
   url = "http://143.198.159.1/ssl/leaders",
   query = list(league = 1, season = 5)
)
}
