% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskScrapers.R
\name{readPosts}
\alias{readPosts}
\title{Scrapes all the posts from a thread page}
\usage{
readPosts(page)
}
\arguments{
\item{page}{A link to a specific forum thread page}
}
\value{
Returns a data frame of all the posts in the thread, by whom and when
}
\description{
Scrapes all the posts from a thread page
}
