% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskScrapers.R
\name{trainingCampLinks}
\alias{trainingCampLinks}
\title{Scrapes the training camp}
\usage{
trainingCampLinks()
}
\value{
Returns a data frame of the predictions
}
\description{
Scrapes the training camp
}
