% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskScrapers.R
\name{activityCheckLinks}
\alias{activityCheckLinks}
\title{Scrapes most recent AC Thread}
\usage{
activityCheckLinks()
}
\value{
Returns a vector of all the AC links
}
\description{
Scrapes most recent AC Thread
}
