% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskScrapers.R
\name{affiliateLinks}
\alias{affiliateLinks}
\title{Scrapes most recent Affiliate thread}
\usage{
affiliateLinks()
}
\value{
Returns a vector of all the Affiliate links
}
\description{
Scrapes most recent Affiliate thread
}
